Tidy prognostic run loop file (#2336)
